Bug fix in SimpleInjectorScope.
The SimpleInjectorScope requested the current execution context scope
during disposal. Not only is this slower, it could in theory lead to
disposing a different scope (or no scope at all) because the user might
have started an inner scope. Instead the created scope is stored in a
private field giving the certainty that it exist and that we dispose
that exact scope.